+++ app/javascript/components/quiz.js
@@ -1,38 +1,41 @@
 // Show next question quiz_questions#index
 
 const displayNextQuestion = () => {
-  const nextBtn = document.getElementById("nextBtn");
-  const endBtn = document.getElementById("endBtn");
-  const backBtn = document.getElementById("backBtn");
+  // const nextBtn = document.getElementById("nextBtn");
+  // const endBtn = document.getElementById("endBtn");
+  // const backBtn = document.getElementById("backBtn");
 
-  let quizBtns = document.querySelectorAll('.quiz-next-btn');
-  let quizContainer = document.querySelectorAll('.quiz-container');
-  // Hide all questions
-  quizContainer.forEach((quiz) => {
-    quiz.classList.add('hide-quiz');
-  });
-  // Show only first one
-  quizContainer[0].classList.remove("hide-quiz");
-
-  quizBtns.forEach((btn) => {
-    btn.addEventListener('click', function (e) {
-      let numb = btn.id.match(/\d/g);
-      numb = numb.join("");
-      numb = parseInt(numb, 10);
-      let containerToHide = document.getElementById(`quiz-container-${numb-1}`);
-      let containerToShow = document.getElementById(`quiz-container-${numb}`);
-
-      containerToHide.classList.add("fade-out");
-
-      setTimeout(function(){
-        containerToHide.classList.add("hide-quiz");
-       }, 800);
-
-      containerToShow.classList.add("fade-in");
-      containerToShow.classList.remove("hide-quiz");
-
+  let quizBtns = document.querySelectorAll(".quiz-next-btn");
+  let quizContainer = document.querySelectorAll(".quiz-container");
+  if (quizContainer.length) {
+    // Hide all questions
+    quizContainer.forEach((quiz) => {
+      quiz.classList.add("hide-quiz");
     });
-  })
+    // Show only first one
+    quizContainer[0].classList.remove("hide-quiz");
+
+    quizBtns.forEach((btn) => {
+      btn.addEventListener("click", function (e) {
+        let numb = btn.id.match(/\d/g);
+        numb = numb.join("");
+        numb = parseInt(numb, 10);
+        let containerToHide = document.getElementById(
+          `quiz-container-${numb - 1}`
+        );
+        let containerToShow = document.getElementById(`quiz-container-${numb}`);
+
+        containerToHide.classList.add("fade-out");
+
+        setTimeout(function () {
+          containerToHide.classList.add("hide-quiz");
+        }, 800);
+
+        containerToShow.classList.add("fade-in");
+        containerToShow.classList.remove("hide-quiz");
+      });
+    });
+  }
 };
 
-export {displayNextQuestion}
+export { displayNextQuestion };
